requirements.txt
# for windows
pywinusb

# for linux
pyusb
